fix plugin uninstall

+++ app/Console/Commands/Plugins/UninstallPluginCommand.php
@@ -3,8 +3,8 @@
 namespace App\Console\Commands\Plugins;
 
 use Illuminate\Console\Command;
-use Illuminate\Support\Facades\Artisan;
 use Illuminate\Support\Facades\File;
+use Illuminate\Support\Facades\Process;
 
 class UninstallPluginCommand extends Command
 {
@@ -24,12 +24,22 @@ class UninstallPluginCommand extends Command
             return;
         }
 
+        $this->info('Removing plugin from Composer...');
+        $result = Process::timeout(0)
+            ->path(base_path())
+            ->run('composer remove '.$this->argument('name'));
+        $this->output->write($result->output());
+
+        if ($result->exitCode() !== 0) {
+            $this->error('Failed to remove plugin from Composer: '.$this->argument('name'));
+
+            return;
+        }
+
         $this->info('Removing plugin files...');
 
         File::deleteDirectory($pluginPath);
 
         $this->info('Plugin '.$this->argument('name').' uninstalled successfully.');
-
-        Artisan::call('plugins:load');
     }
 }
